# SPU: Evaluated Thread-blocked LR messaging