--- a/rpcs3/rpcs3qt/emu_settings_type.h
+++ b/rpcs3/rpcs3qt/emu_settings_type.h
@@ -52,6 +52,7 @@ enum class emu_settings_type
 	MaxPreemptCount,
 	SPUProfiler,
 	DisableSpinOptimization,
+	EnabledSPUEventsBusyLoop,
 
 	// Graphics
 	Renderer,
@@ -250,6 +251,7 @@ inline static const std::map<emu_settings_type, cfg_location> settings_location 
 	{ emu_settings_type::MaxPreemptCount,          { "Core", "Max CPU Preempt Count"}},
 	{ emu_settings_type::SPUProfiler,              { "Core", "SPU Profiler"}},
 	{ emu_settings_type::DisableSpinOptimization,  { "Core", "Disable SPU GETLLAR Spin Optimization"}},
+	{ emu_settings_type::EnabledSPUEventsBusyLoop, { "Core", "SPU Reservation Busy Waiting Enabled"}},
 
 	// Graphics Tab
 	{ emu_settings_type::Renderer,                   { "Video", "Renderer"}},