… rpcs3/rpcs3qt/tooltips.h