* In Metal Gear Solid 4, SPU_RdEventStat event polling seems to be lightspeed fast for some SPU threads but not for others.
I have implemented an algorithm that evlautes the likelyhood of need of SPU_RdEventStat to respond at CPU core speed based on previous polling history.

For example for its "MGS4_URGENTCellSpursKernel0" SPU thread, SPU polling time was rarely ever above 100-microseconds.

### This behaviour is enabled when enabling the setting "SPU Events Busy Loop" in Advanced tab.

Performance Test Results on Ryzen 8945HS by @cipherxof :

**115 fps -> 130 fps (13% improvement)**

Performance Test Results on i7 12700k by @cipherxof :

**160 fps -> 190 fps (18% improvement)**

Before:
![image](https://github.com/user-attachments/assets/3e261d5d-e3a2-4153-ab7d-fc7dc1e6bda7)

After:
![image](https://github.com/user-attachments/assets/ccaa31e5-0e26-4eb2-bc43-f92950190692)

Please test CPU usage difference too.

**At the moment: Only 12+ threaded CPUs have this feature enabled in this build!**