@elad335 2025-03-21
@cipherxof Updated the algorithm, can you retest?