@cipherxof 2025-03-21
12700k linux. This PR has some bad frame spikes into the 170s, but the highs are higher.

Master:

![Screenshot From 2025-03-21 11-17-25](https://github.com/user-attachments/assets/502f82e7-a26f-4d63-bd64-1c95776825ad)

PR:

![Screenshot From 2025-03-21 11-14-05](https://github.com/user-attachments/assets/addcc4b5-2410-4d4f-aea1-0caeaba967fa)