@Megamouse 2025-03-21
> 12700k master:
> 
> ![Screenshot 2025-03-21 074846](https://private-user-images.githubusercontent.com/5994581/425597392-fc3390fa-eef5-4103-be1c-9a624c58644a.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDI1ODE4MTEsIm5iZiI6MTc0MjU4MTUxMSwicGF0aCI6Ii81OTk0NTgxLzQyNTU5NzM5Mi1mYzMzOTBmYS1lZWY1LTQxMDMtYmUxYy05YTYyNGM1ODY0NGEucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDMyMSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTAzMjFUMTgyNTExWiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9YWUwODk0YzRiOTAyZDg1YjA1ODU3YmUwYTUyMTk2ZDc5NzAyMWQ1MzM1NjMyNTBiYWYwOGE0OTQxN2ZiMDdiMiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.WaUrrxr_NSh-4tjxC0P78eLFVNXMs71d0N6Pk8f5xQI)
> 
> 12700k pr:
> 
> ![Screenshot 2025-03-21 175223](https://private-user-images.githubusercontent.com/5994581/425597439-155e0770-8abe-4f97-85ad-7c637c231ff7.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDI1ODE4MTEsIm5iZiI6MTc0MjU4MTUxMSwicGF0aCI6Ii81OTk0NTgxLzQyNTU5NzQzOS0xNTVlMDc3MC04YWJlLTRmOTctODVhZC03YzYzN2MyMzFmZjcucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDMyMSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTAzMjFUMTgyNTExWiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9MjM1YTkwMDY5N2Q2YWU0Mzk5YjRmMDc4N2U4ZTFhZDU2ZDk5MmQ2ZGQzM2Y4MWUyZGYxZmVjYjI4YWY4MmQzMyZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.gKWdfbdRy2BYJW2GHJynXzj2sLIAd2a31JVcreO1mkI)

Keep in mind that this master build doesn't have the LTO stuff yet.
So better retry this comparison with actual master