@Jonathan44062 2025-03-21
Performance regression with SU
Master
![image](https://github.com/user-attachments/assets/00a665da-9866-49e1-9383-985f43ed1541)

PR
![image](https://github.com/user-attachments/assets/3e608866-ec4f-4030-8630-e2cc233602f3)