@elad335 2025-03-22
@cipherxof @Jonathan44062 I've pushed a bugfix for performance, please retest.