@cipherxof 2025-03-22
> @cipherxof @Jonathan44062 I've pushed a bugfix for performance, please retest.

I seem to have lost all of the performance gains on Windows.

<details>
  <summary>Screenshots</summary>
  
![Untitled](https://github.com/user-attachments/assets/aa056eb0-896d-49d0-9606-160870d7f957)

![Screenshot From 2025-03-22 01-49-22](https://github.com/user-attachments/assets/51c4dbe3-9b7b-40d4-8c13-978aab66184c)

</details>