@elad335 2025-03-22
This pull request is much fun so far.
So, I had a slight suspicioun that the reason the PPU managing the SPU with the URGENT jobqueue is so eager to push a new command is because of the MGS4 stability hack which drains the jobqueue of two places out of four. So the SPU can have in theory onl;y half the work it can before it finishes its execution and goes to SPU_RdEventStat polling state.
Meanwhile, on the PPU side, there is no syscall to explcicitly join the jobqueue, only to spin on the function that tries to add a new urgent command until it succeeds. Which explains why PPU response time is so fast!

I had yet another suspsicioin that the problem of all of this is the SPU reset which should be atomic and is not in RPCS3 because it is probably based in a MFC_PUT_CMD command. I added a specific case of atomicity that targets SPURS urgent jobqueue reset. 
Seems to work!

In the end, we may both increase the performance of MGS4 and fix its freeze.

**@cipherxof Test this pull request WITHOUT THE FREEZE PATCH!**