@cipherxof 2025-03-22
Performance is still not restored for me on Windows either.