@elad335 2025-05-06
@cipherxof the old method was turned out to be wrong yet beneficial for MGS4 for high threaded CPUs, so instead I turned it to a setting. Named "SPU Reservation Busy Waiting Enabled".
Default is disabled.