@cipherxof 2025-05-06
Oddly enough, I am not seeing much if any difference with the setting enabled or disabled. I will re-test my 12700k later. Ignore the 1% lows as they're likely from alt-tabbing or screenshotting.

Enabled:
![Screenshot 2025-05-06 120417](https://github.com/user-attachments/assets/6ae50908-d1c5-480c-94c4-5a6500f9ee89)

Disabled:
![Screenshot 2025-05-06 120645](https://github.com/user-attachments/assets/dd454298-b530-4810-9c35-9b0cb6d4bdf6)