@cipherxof 2025-05-06
And here is the master RPCS3 build. So  this PR is definitely improving the performance, but the setting doesn't seem to make a difference.

![image](https://github.com/user-attachments/assets/c3045bb0-fa12-4606-a0ae-6d40d6ed5ee8)