@elad335 2025-05-06
> And here is the master RPCS3 build. So this PR is definitely improving the performance, but the setting doesn't seem to make a difference.
> 
> ![image](https://private-user-images.githubusercontent.com/5994581/440925855-c3045bb0-fa12-4606-a0ae-6d40d6ed5ee8.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDY1NTk4NDksIm5iZiI6MTc0NjU1OTU0OSwicGF0aCI6Ii81OTk0NTgxLzQ0MDkyNTg1NS1jMzA0NWJiMC1mYTEyLTQ2MDYtYTBhZS02ZDQwZDZlZDVlZTgucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDUwNiUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTA1MDZUMTkyNTQ5WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9ZjYyOGY5MmI1NDJhNjQ3Y2IzYTMwZWVhMDBmMmMzNjcyNTc3Njk2YjlmMTNmYmQzYzEyYzI2NzQxZjE0OGM1NyZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.2mm_RfNz8nJhTAHWMWp6QTj11LBpTjaLut9RwyTRFc4)

Try to enable "Disable SPU GETLLAR Spin Optimization" with the PR build amd master