@cipherxof 2025-05-06
@elad335 

**PR:**

SPU Reservation Busy Waiting Enabled & Enable SPU GETLLAR Spin Optimization: 129fps
SPU Reservation Busy Waiting Enabled & Disable SPU GETLLAR Spin Optimization: 120fps
SPU Reservation Busy Waiting Disabled & Disable SPU GETLLAR Spin Optimization: 110fps

**Master:**

Enable SPU GETLLAR Spin Optimization: 117fps
Disable SPU GETLLAR Spin Optimization: 113fps