@cipherxof 2025-05-06
Re-tested my 12700k on Linux, and the bad frame spikes are gone. Performance is better than master as well.

<details>
  <summary>Master</summary>
  
![Screenshot From 2025-05-06 14-25-21](https://github.com/user-attachments/assets/fa13d757-5175-463b-a270-5e152b8e8402)
  
</details>

<details>
  <summary>PR</summary>
  
![Screenshot From 2025-05-06 14-29-28](https://github.com/user-attachments/assets/1c1bcc65-163b-4e2c-89ba-9b1959fb4475)
  
</details>