@elad335 2025-05-08
Added it as a setting: "SPU Events Busy Loop" in Advanced tab.