@chuwak86 2025-05-15
Since this build I have huge audio stators and frame drops in TLOU
windows 10, intel 9600k, rtx3070

downgrade to 17924 solves stutters, 

disabling spin optimization does not solve stutters