@digant73 2025-08-20
> Since this build I have huge audio stutters and frame drops in TLOU windows 10, intel 9600k, rtx3070
> 
> downgrade to 17924 solves stutters,
> 
> disabling spin optimization does not solve stutters

if you are still here, your issue should be fixed now by #17437 just merged on new release 0.0.37-18108.
A feedback on that is welcome