@chuwak86 2025-08-20
> optimizati
Yes, Im here
yes, latest builds does not have this problem